Refactoring and improvements
