+++ api/service_plan_controller.go
@@ -59,7 +59,7 @@ func (c *ServicePlanController) Routes() []web.Route {
 				Method: http.MethodPatch,
 				Path:   fmt.Sprintf("%s/{%s}", web.ServicePlansURL, PathParamID),
 			},
-			Handler: c.PatchLabelsOnly,
+			Handler: c.PatchObject,
 		},
 	}
 }
